Proof of refinement triple for lock free read. Still need to update recovery proof.
The abstraction relation had to be updated to account for the fact that readers
may have to do a form of helping in the simulation proof.  Consider a scenario
where a write finishes the write to d0, and is then paused. Suppose then that
d1 fails. A read begins and first reads d1, finding that it has failed, so it
will then read and return the value in d0. In other words, because d1 has
failed, the write has effectively already taken place, so its linearization
point needs to be before the read. Thus, in this case, the reader will first
simulate the step of the writer and then simulate its own read.