Define semantically typed translation of a layer, start proofs.
Translations are indexed by a notion of a "semantic" typed translation.
These semantic types consist of a Kripke structure, where the forcing
relation on worlds relates values in the abstract layer to concrete
values.

The rule for compiling (x <- m; f(x)) says that the translation of f has to be
valid under the assumption that we're in a world where the return value x from
m is well-related to the return value of its translation.

The goal of this set-up is to avoid too much bureaucratic deep-embedding
of a type system.