HOCAP spec for log reset. Had to strengthen sequential spec.
Re-proving this required strengthening the underlying "sequential" reset spec
to the style where the precondition is a conjunction of the crash condition and
a fupd to run at the durability point (like for wpc_Write_fupd).

The issue is that when we wrap this sequential code with a lock and then try to
prove a HOCAP spec on top, in the event of a crash, we either have to execute
the HOCAP fupd or not, based on whether the log was reset or not.

Before the changes on this branch, one could run an arbitrary fupd when proving
the crash condition but now you cannot.

I don't think this is a problem since all of our modules should be exporting
HOCAP style specs anyway, and the challenge of taking some totally sequential code
and making it concurrent without having to improve the sequential spec at all
was kind of an artificial challenge.